+++ libs/features/invocation-route/src/lib/entries/Input.tsx
@@ -1,6 +1,6 @@
 import { InputJournalEntryType } from '@restate/data-access/admin-api';
 import { EntryProps } from './types';
-import { Expression, InputOutput } from '../Handler';
+import { Expression, InputOutput } from '../Expression';
 import { Headers } from '../Headers';
 import { Value } from '../Value';
